backend/language_proj/language_proj/urls.py


from django.conf.urls.static import static
from django.conf import settings
from django.contrib import admin
from django.urls import path, include
from rest_framework_jwt.views import obtain_jwt_token


urlpatterns = [
    path('admin/', admin.site.urls),
    path('api/', include('language_app.urls')),
    path('token-auth/', obtain_jwt_token),
    path('core/', include('core.urls'))

]

if settings.DEBUG:
        urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
